# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Assignment": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "video_id": {
            "minLength": 1,
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "video_id",
          "at"
        ],
        "title": "Assignment",
        "type": "object"
      },
      "AutoScheduleRequest": {
        "properties": {
          "horizon_days": {
            "default": 28,
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Horizon Days",
            "type": "integer"
          },
          "long_per_week": {
            "default": 1,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Long Per Week",
            "type": "integer"
          },
          "shorts_per_week": {
            "default": 3,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Shorts Per Week",
            "type": "integer"
          },
          "videos": {
            "items": {
              "$ref": "#/components/schemas/PendingVideo"
            },
            "title": "Videos",
            "type": "array"
          }
        },
        "required": [
          "videos"
        ],
        "title": "AutoScheduleRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PendingVideo": {
        "description": "One video waiting for a slot.\n\nWas `dict`, which meant a missing `id` reached `auto_schedule` and came back\nas a 500 with a KeyError \u2014 four ordinary malformed payloads all did. Declaring\nthe shape turns each of them into a 422 naming the field.",
        "properties": {
          "format": {
            "default": "short",
            "enum": [
              "short",
              "long"
            ],
            "title": "Format",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "ready_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ready At"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "PendingVideo",
        "type": "object"
      },
      "PublishRequest": {
        "description": "Choices the operator makes at the approval gate.\n\nAll optional: the defaults publish the top-scored title and first thumbnail\nimmediately and publicly, which is the common case.",
        "properties": {
          "chosen_thumbnail_index": {
            "default": 0,
            "minimum": 0.0,
            "title": "Chosen Thumbnail Index",
            "type": "integer"
          },
          "chosen_title_index": {
            "default": 0,
            "minimum": 0.0,
            "title": "Chosen Title Index",
            "type": "integer"
          },
          "made_for_kids": {
            "default": false,
            "title": "Made For Kids",
            "type": "boolean"
          },
          "playlist_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playlist Id"
          },
          "privacy": {
            "default": "public",
            "pattern": "^(public|unlisted|private)$",
            "title": "Privacy",
            "type": "string"
          },
          "publish_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish At"
          }
        },
        "title": "PublishRequest",
        "type": "object"
      },
...
      "ScheduleRequest": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "video_id",
          "at"
        ],
        "title": "ScheduleRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "engine__api__publishing__ApplyRequest": {
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/Assignment"
            },
            "title": "Assignments",
            "type": "array"
          }
        },
        "required": [
          "assignments"
        ],
        "title": "ApplyRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Studio Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/google/callback": {
      "get": {
        "operationId": "finish_auth_v1_auth_google_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finish Auth",
        "tags": [
          "publishing"
        ]
      }
    },
...
    "/v1/calendar/auto": {
      "post": {
        "description": "Fill the calendar automatically.\n\nReturns a *plan*, and does not apply it. Scheduling a month of uploads is exactly\nthe kind of thing that should be reviewed before it happens.",
        "operationId": "auto_v1_calendar_auto_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Auto V1 Calendar Auto Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Auto",
        "tags": [
          "publishing"
        ]
      }
    },
    "/v1/calendar/auto/apply": {
      "post": {
        "description": "Book a whole plan, or as much of it as the rules allow.\n\nTwo things were wrong with the old version, and they compounded.\n\nIt **wrote as it went**, so a malformed entry at position three left one and\ntwo booked and the caller with a 500 \u2014 no way to know what had landed short\nof re-reading the calendar. Everything is now checked before the first write,\nso the outcome is all-or-reported, never half-applied-and-unexplained.\n\nAnd it **never called `validate_move`**, which `schedule_one` twenty lines\nabove does on every manual drag. The same times that endpoint 409s \u2014 in the\npast, too close to another upload, over the day's quota \u2014 were persisted here\nwithout complaint, which is how a calendar ends up describing a schedule\nYouTube will refuse.\n\nRejected entries come back in `skipped` with a reason rather than failing the\nrequest: a fourteen-video plan with one bad slot should book thirteen and say\nwhich one it did not.",
        "operationId": "apply_plan_v1_calendar_auto_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/engine__api__publishing__ApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Apply Plan V1 Calendar Auto Apply Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Plan",
        "tags": [
          "publishing"
        ]
      }
    },
    "/v1/calendar/schedule": {
      "post": {
        "description": "A manual drag. Validated, and warned about even when permitted.",
        "operationId": "schedule_one_v1_calendar_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Schedule One V1 Calendar Schedule Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedule One",
        "tags": [
          "publishing"
        ]
      }
    },
    "/v1/calendar/schedule/{video_id}": {
      "delete": {
        "operationId": "unschedule_v1_calendar_schedule__video_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Unschedule V1 Calendar Schedule  Video Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unschedule",
        "tags": [
          "publishing"
        ]
      }
    },
    "/v1/calendar/slots": {
      "get": {
        "description": "Ranked publish times, so the calendar can show *why* a slot is good.",
        "operationId": "slots_v1_calendar_slots_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 14,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Slots V1 Calendar Slots Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Slots",
        "tags": [
          "publishing"
        ]
      }
    },
...
    "/v1/jobs/{job_id}/publish": {
      "post": {
        "description": "Publish a finished video. **This is the approval gate.**\n\n`CLAUDE.md` non-negotiable #3: nothing publishes without an explicit approval\ngate. That gate is this endpoint \u2014 it is never a stage of the `video` workflow,\nbecause a workflow that publishes as its last step has no gate at all.\n\nAuto-publish (`automation.py`) calls the same function, so manual and unattended\npublishing share one code path and one set of blockers. Skipping the checks is\nnot offered: a series with `auto_publish=True` skips the *waiting*, not the\n*checks*.",
        "operationId": "publish_job_v1_jobs__job_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Publish Job V1 Jobs  Job Id  Publish Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Job"
      }
    },
...
  }
}